+++ src/settings.ts
@@ -9,5 +9,5 @@ export const PROTECT_API_ERROR_LIMIT = 10;
 // Interval, in seconds, to wait before trying to access the API again once we've hit the PROTECT_API_ERROR_LIMIT threshold.
 export const PROTECT_API_RETRY_INTERVAL = 300;
 
-// Protect API response timeout, in seconds. This should never be greater than 5 seconds.
-export const PROTECT_API_TIMEOUT = 3.5;
+// Protect API response timeout, in milliseconds. This should never be greater than 5 seconds.
+export const PROTECT_API_TIMEOUT = 4000;